bots: Put system bots in each realm.
- migration to add the bots to every existing realm
- code changes to make creation of these bots part of the real creation
  process. This is done via the setup_realm_internal_bots function that
  already existed, but did nothing, because REALM_INTERNAL_BOTS was
  empty. We make these system bots REALM_INTERNAL_BOTS. Nagios bots are
  skipped, as they're not cross-realm bots, they just send messages
  within their zulipinternal realm.
- get_system_bots and friends are adjusted to take a realm argument to
  know from which realm to grab the bot. This makes the functions
  get_user_by_id_in_realm_including_cross_realm and
  get_user_including_cross_realm a bit nonsensical, but we just want to
  make things work for now. Redundant code like this will be cleaned up
  later.

This commit breaks a lot of things, it just lays the foundation for
follow-up changes. All calls to get_system_bot and friends need to now
be adjusted, and various other tweaks will be needed. These will be
addressed in the next commits.
