+++ zerver/lib/cache.py
@@ -499,8 +499,8 @@ def user_profile_delivery_email_cache_key(delivery_email: str, realm: "Realm") -
     return f"user_profile_by_delivery_email:{make_safe_digest(delivery_email.strip())}:{realm.id}"
 
 
-def bot_profile_cache_key(email: str, realm_id: Optional[int] = None) -> str:
-    return f"bot_profile:{make_safe_digest(email.strip())}"
+def bot_profile_cache_key(email: str, realm_id: int) -> str:
+    return f"bot_profile:{make_safe_digest(email.strip())}:{realm_id}"
 
 
 def user_profile_by_id_cache_key(user_profile_id: int) -> str:
@@ -592,7 +592,7 @@ def delete_user_profile_caches(user_profiles: Iterable["UserProfile"]) -> None:
         )
         if user_profile.is_bot and is_cross_realm_bot_email(user_profile.email):
             # Handle clearing system bots from their special cache.
-            keys.append(bot_profile_cache_key(user_profile.email))
+            keys.append(bot_profile_cache_key(user_profile.email, user_profile.realm_id))
 
     cache_delete_many(keys)
