+++ zerver/models.py
@@ -3092,38 +3092,23 @@ def get_active_user_profile_by_id_in_realm(uid: int, realm: Realm) -> UserProfil
 def get_user_including_cross_realm(email: str, realm: Realm) -> UserProfile:
     if is_cross_realm_bot_email(email):
         return get_system_bot(email, realm.id)
-    assert realm is not None
     return get_user(email, realm)
 
 
 @cache_with_key(bot_profile_cache_key, timeout=3600 * 24 * 7)
 def get_system_bot(email: str, realm_id: int) -> UserProfile:
-    """
-    This function doesn't use the realm_id argument yet, but requires
-    passing it as preparation for adding system bots to each realm instead
-    of having them all in a separate system bot realm.
-    If you're calling this function, use the id of the realm in which the system
-    bot will be after that migration. If the bot is supposed to send a message,
-    the same realm as the one *to* which the message will be sent should be used - because
-    cross-realm messages will be eliminated as part of the migration.
-    """
-    return UserProfile.objects.select_related().get(email__iexact=email.strip())
+    return UserProfile.objects.select_related().get(
+        email__iexact=email.strip(), realm_id=realm_id, is_bot=True
+    )
 
 
 def get_user_by_id_in_realm_including_cross_realm(
     uid: int,
-    realm: Optional[Realm],
+    realm: Realm,
 ) -> UserProfile:
     user_profile = get_user_profile_by_id(uid)
-    if user_profile.realm == realm:
-        return user_profile
 
-    # Note: This doesn't validate whether the `realm` passed in is
-    # None/invalid for the CROSS_REALM_BOT_EMAILS case.
-    if user_profile.delivery_email in settings.CROSS_REALM_BOT_EMAILS:
-        return user_profile
-
-    raise UserProfile.DoesNotExist()
+    return user_profile
 
 
 @cache_with_key(realm_user_dicts_cache_key, timeout=3600 * 24 * 7)
